requirements.txt
fastapi
uvicorn
pillow
ultralytics>=8.0
torch>=2.0
